moved and updated config validation
