+++ receiver/postgresqlreceiver/factory.go
@@ -4,9 +4,6 @@ package postgresqlreceiver
 
 import (
 	"context"
-	"fmt"
-	"net/url"
-	"regexp"
 	"time"
 
 	"go.opentelemetry.io/collector/component"
@@ -14,7 +11,6 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver/receiverhelper"
 	"go.opentelemetry.io/collector/receiver/scraperhelper"
-	"go.uber.org/multierr"
 )
 
 const (
@@ -57,27 +53,4 @@ func createMetricsReceiver(
 		&cfg.ScraperControllerSettings, params.Logger, consumer,
 		scraperhelper.AddScraper(scraper),
 	)
-}
-
-func (cfg *Config) Validate() error {
-	var errs []error
-	var validPort = regexp.MustCompile(`(:\d+)`)
-
-	if cfg.Username == "" {
-		errs = append(errs, fmt.Errorf("missing required field 'username'"))
-	}
-	if cfg.Password == "" {
-		errs = append(errs, fmt.Errorf("missing required field 'password'"))
-	}
-	if cfg.Database == "" {
-		errs = append(errs, fmt.Errorf("missing required field 'database'"))
-	}
-	if cfg.Endpoint == "" {
-		errs = append(errs, fmt.Errorf("missing required field 'endpoint'"))
-	} else if _, err := url.Parse(cfg.Endpoint); err != nil {
-		errs = append(errs, fmt.Errorf("invalid url specified in field 'endpoint'"))
-	} else if !validPort.MatchString(cfg.Endpoint) {
-		errs = append(errs, fmt.Errorf("invalid port specified in field 'endpoint'"))
-	}
-	return multierr.Combine(errs...)
 }
